static-pools: slight refactoring and renaming
- Drop flags.go and move stuff over to config.go as we really don't have
  any flags, anymore
- Rename type conf to config
- Rename global var cfg to conf. This hopefully makes the code easier to
  follow as many functions/methods also have an argument named cfg.
